Fix cases when git repo is cleaned/just cloned
Add placeholder file to keep directory.
Check sha of zip to see if directory is empty (Not
sure if that is really working properly. I did run
this twice so that terraform showed changes. Could it be
that on first run the that the zip that was sent as layer
was actually empty?)

Signed-off-by: Rami Rahikkala <rami.rahikkala@fluentia.fi>